Split tests into different test classes
