+++ test-projects/basic/allshared/build.gradle.kts
@@ -7,6 +7,7 @@ plugins {
 
 kotlin {
     listOf(
+        iosX64(),
         iosArm64(),
         iosSimulatorArm64()
     ).forEach {
